Working intel nic ISR, w00t! Still needs lots of work

+++ src/scratch/pci-scanner/types.h
@@ -13,7 +13,7 @@ typedef long long		int64_t;
 
 typedef	unsigned char		uint8_t;
 typedef	unsigned short		uint16_t;
-typedef unsigned long		uint32_t;
+typedef unsigned int		uint32_t;
 typedef unsigned long long	uint64_t;
 
 typedef uint32_t		uint_t;
